sp fixtures
